backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
pytesseract==0.3.10
Pillow>=10.2.0
PyPDF2==3.0.1
python-dateutil==2.8.2
regex==2024.9.11
fuzzywuzzy==0.18.0
rapidfuzz==3.9.6
Flask-CORS==4.0.0
pandas>=2.0.0
PyJWT==2.8.0
bcrypt==4.1.1
passlib==1.7.4
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
gunicorn==21.2.0
python-dotenv==1.0.0
# python-Levenshtein==0.21.1   # removed (forces compilation on Windows, rapidfuzz is a drop-in replacement)
